Adjust Model Datatypes
-Swap int for long
-Add MaxLength to Counterparty name
-Update validator to use correct min/max values

+++ FinanceApp.Client/Classes/FinanceClient.cs
@@ -70,7 +70,7 @@ public class FinanceClient : IClient
 					{
 						CreateTransaction transaction = new()
 						{
-                            Value = new RequestField<int>
+                            Value = new RequestField<long>
                             {
                                 Value = value
                             },
